Add initial tests
